Fix some more C code exhibiting "undefined behaviour"
These were found with the -fsanitize=undefined option to clang.

src/cmd/ksh93/edit/history.c,
src/lib/libast/sfio/sf{rd,seek,setbuf}.c:
- Fix four cases of "adding a zero offset to a null pointer" by
  checking that the pointer is non-null first.

src/cmd/ksh93/sh/init.c: nv_mapchar():
- When an 'np' of NULL is passed to this function, it will return
  with either a NULL pointer or a pointer with a value of 1, which
  is meant to be used as a boolean (see bltins/typeset.c line 351).
  Instead of ((Namfun_t*)0) + 1, which is undefined behaviour,
  return ((Namfun_t*)1), which is not.

src/cmd/ksh93/sh/arith.c: arith():
- It is undefined behaviour to typecast from a float/double (r) to
  an integer type if the value is not within the representable
  range of the integer type. However, typecasting from a larger to
  a smaller integer type is "implementation-defined behaviour" in
  that case, which is much less bad. So typecast the double to
  intmax_t instead, and then typecast from intmax_t to int16_t or
  int32_t.

+++ src/cmd/ksh93/sh/arith.c
@@ -279,9 +279,9 @@ static Sfdouble_t arith(const char **ptr, struct lval *lvalue, int type, Sfdoubl
 		else if((attr & NV_INT64)==NV_INT64)		/* long signed integer */
 			r = (intmax_t)n;
 		else if((attr & NV_INT16)==NV_INT16)		/* short signed integer */
-			r = (int16_t)n;
+			r = (int16_t)((intmax_t)n);
 		else if((attr & NV_INT32)==NV_INT32)		/* normal signed integer */
-			r = (int32_t)n;
+			r = (int32_t)((intmax_t)n);
 #if _AST_release
 		else	r = n;					/* should never happen */
 #else
